Started fixing the code, not fully done
