+++ public/js/FrontWeaponStuff/WeaponDrawning.js
@@ -69,3 +69,4 @@ class WeaponDrawing {
         }
     }
 }
+
